workflows: add one-line-cr-bot
We want fully automated code analysis. This tool and template make this
analysis possible with cppcheck and infer.

New defects will be raised, and the step will fail if there are new
defects.

Please note, this is currently an PoC, and might fail once in a while.
Furthermore, the tools can produce false positives, so that this checker
might want to be ignored once in a while.

Signed-off-by: Norbert Manthey <nmanthey@conp-solutions.com>