backtrack: add trail saving
This change implements saving the trail during backtracking to be able
to reuse it during unit propagation again, instead of visiting watch
lists. The idea has been described in "Trail Saving on Backtrack" in
the SAT conference 2020 by Randy Hickey and Fahiem Bacchus.

The implementation is close to the one in the 2020 solver
Maple_LCM_Dist-alluip-trail, except during backtracking. There, we do
not preallocate the saved trail first, but instead collect all literals
that will be unassigned in the stack first, and next reverse the stack.
This allows integration with non-chronological backtracking.

Signed-off-by: Norbert Manthey <nmanthey@conp-solutions.com>