analysis: sort learnt clause by level
Literals in a learned clause are added in the order they are spotted
during learning. To focus more on potentially unassigned literals, move
the literals from lower decision levels to the back of the clause.

To avoid spending too much time on very long learned clauses, that we
will likely remove soon, introduce a parameter that controls which
clauses to sort.

Signed-off-by: Norbert Manthey <nmanthey@conp-solutions.com>