lcm: drop duplicates
During LCM, we hit many duplicate clauses. There are proposed solutions
to memorize all learnt clauses to detect duplicates.

This change proposes to store only a small set of most recently clauses
that have been the result of simplification, and use this set to check
whether newly simplified clauses match the recent ones.

In case we detect a duplicate, we drop the current clause and remove it
from the solvers structures.

Signed-off-by: Norbert Manthey <nmanthey@conp-solutions.com>