removeClause: allow to not consider proof
In some cases, we might remove a clause from the solver structure,
but not from the proof. Such a case is LCM reducing a clause, and
next ignoring the reduced clause, as it is a duplicate of another
clause.

This change extends the removeClause interface to be able to specify
whether the proof should be considered. By default, the proof is
considered.

Signed-off-by: Norbert Manthey <nmanthey@conp-solutions.com>