parallel: add initial sync mechanism
This commit adds a stub to sync the solver execution. Essentially,
the method marks the areas in which each solver should update
the shared structures.

Note: termination of search is still not controlled, and can lead to
crashed, e.g. in case a thread is still running while the main thread
already deletes the solver object thanks for a fast primary solver.

Signed-off-by: Norbert Manthey <nmanthey@conp-solutions.com>