compare: extend to parallel solver
As the script allows to compare solver behavior, we can also use it to
check whether the parallel solver is deterministic, i.e. it finishes
solving a given input formula with the same stats when being run
multiple times.

This change adds a few patterns to extract behavior from the parallel
solver.

Signed-off-by: Norbert Manthey <nmanthey@conp-solutions.com>